+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousPhoneTransfersRepository.java
@@ -4,6 +4,8 @@ import com.bank.antifraud.entity.SuspiciousPhoneTransfersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface SuspiciousPhoneTransfersRepository extends JpaRepository<SuspiciousPhoneTransfersEntity, Long> {
+public interface SuspiciousPhoneTransfersRepository extends JpaRepository<SuspiciousPhoneTransfersEntity, BigInteger> {
 }
