+++ antifraud/src/main/java/com/bank/antifraud/service/SuspiciousAccountTransfersService.java
@@ -1,17 +1,15 @@
 package com.bank.antifraud.service;
 
-import com.bank.antifraud.dto.AuditDto;
 import com.bank.antifraud.dto.SuspiciousAccountTransfersDto;
-import com.bank.antifraud.entity.AuditEntity;
 import com.bank.antifraud.entity.SuspiciousAccountTransfersEntity;
 
+import java.math.BigInteger;
 import java.util.List;
-import java.util.Optional;
 
 public interface SuspiciousAccountTransfersService {
     public SuspiciousAccountTransfersEntity save(SuspiciousAccountTransfersDto suspiciousAccountTransfersDto);
-    public SuspiciousAccountTransfersEntity findById(Long id);
+    public SuspiciousAccountTransfersEntity findById(BigInteger id);
     public List<SuspiciousAccountTransfersEntity> findAll();
-    public void delete(Long id);
+    public void delete(BigInteger id);
     public SuspiciousAccountTransfersEntity update(SuspiciousAccountTransfersDto suspiciousAccountTransfersDto);
 }
